Migrate additional methods to OutputWriter and add fln() method
This commit continues the OutputWriter refactoring by migrating more methods from TestCaseRun to the base class and adding the fln() convenience method.

## New Method: fln()

Added `fln(text)` as a convenience method for writing failed text with newline:
- Built on f() primitive
- Similar to tln() and iln() but for failed output
- Does NOT call fail() (unlike failln() which marks the line as failed)

**Difference between fln() and failln():**
- `fln()`: Write text as failed tokens with newline (doesn't mark line as failed)
- `failln()`: Write text as failed AND mark entire line as failed (causes test failure)

## Migrated Methods to OutputWriter

The following methods were moved from TestCaseRun to OutputWriter to make them available in all output contexts:

### Timing Methods
- `tmsln(f, max_ms)`: Run function and measure time, fail if exceeds max_ms
- `imsln(f)`: Run function and measure time (info, never fails)

### Image Methods
- `timage(file, alt_text)`: Add markdown image in test stream (tested)
- `iimage(file, alt_text)`: Add markdown image in info stream (not tested)

### List Methods
- `tlist(list, prefix)`: Write list to test stream
- `ilist(list, prefix)`: Write list to info stream
- `tset(items, prefix)`: Write set with out-of-order comparison

### Testing Methods
- `must_apply(it, title, cond, error_message)`: Assertion with decoration
- `must_contain(it, member)`: Assert membership
- `must_equal(it, value)`: Assert equality
- `must_be_a(it, typ)`: Assert type
- `it(name, it)`: Create TestIt wrapper for fluent assertions
- `tformat(value)`: Convert value to JSON format structure

### Other Methods
- `header(header)`: Create header line that operates as anchor

## Refactoring: TestIt and value_format

Created new file `booktest/reporting/testing.py` and moved:
- `TestIt` class
- `value_format` function

These were previously in testcaserun.py but are now in a dedicated testing utilities module. Updated imports in __init__.py accordingly.

## Bug Fixes

Fixed token coloring issues in testcaserun.py:
- Fixed logic error: changed `len(self.line_markers) > 0` to `== 0`
- Improved marker sorting: now sorts by both position and type `(m[0], m[2])`
- Fixed token extraction to prevent overlaps: `line[max(last_pos,start_pos):end_pos]`

## Version Bump

Bumped version to 1.0.5 in pyproject.toml.

## Testing

All 132 tests pass successfully.
